change: Change structure from React to Next
During conversations between the members we decided
to use next.js instead of pure react, this was a choice made because the members were interested in learning and it was what made the most sense at the time of migration.
Furthermore, the recommendation from the react website is to use a framework to speed up the project.
Refs: #3

Signed-off-by: Kawhan <kawhanlaurindo@outlook.com>